revert: undo test restructure commits (moving to separate PR)
