+++ compute_space/compute_space/__init__.py
@@ -1,1 +1,4 @@
+from pathlib import Path
 
+COMPUTE_SPACE_PACKAGE_DIR = Path(__file__).resolve().parent.parent
+OPENHOST_PROJECT_DIR = COMPUTE_SPACE_PACKAGE_DIR.parent
