+++ compute_space/compute_space/config.py
@@ -10,9 +10,6 @@ import tomli_w
 import typed_settings
 from quart import current_app
 
-COMPUTE_SPACE_PACKAGE_DIR = Path(__file__).resolve().parent.parent
-OPENHOST_PROJECT_DIR = COMPUTE_SPACE_PACKAGE_DIR.parent
-
 
 @attr.s(auto_attribs=True, frozen=True)
 class Config:
