+++ compute_space/tests/test_integration.py
@@ -22,17 +22,18 @@ from pathlib import Path
 import jwt as pyjwt
 import pytest
 import requests
-from container import container_cleanup
 from loguru import logger
 
-from compute_space.config import OPENHOST_PROJECT_DIR
+from compute_space import OPENHOST_PROJECT_DIR
 from compute_space.core.caddy import generate_caddyfile
 from compute_space.core.data import provision_data
 from compute_space.core.manifest import AppManifest
 from compute_space.testing import wait_app_running
-from conftest import _make_config_and_env
-from conftest import _start_router_process
-from conftest import _stop_router_process
+
+from .conftest import _make_config_and_env
+from .conftest import _start_router_process
+from .conftest import _stop_router_process
+from .container import container_cleanup
 
 _APPS_DIR = str(OPENHOST_PROJECT_DIR / "apps")
