+++ self_host_cli/doctor.py
@@ -14,7 +14,7 @@ import socket
 import subprocess
 import sys
 
-from compute_space.config import COMPUTE_SPACE_PACKAGE_DIR
+from compute_space import COMPUTE_SPACE_PACKAGE_DIR
 
 
 class _Check:
