+++ self_host_cli/up.py
@@ -12,7 +12,7 @@ import socket
 import subprocess
 import sys
 
-from compute_space.config import COMPUTE_SPACE_PACKAGE_DIR
+from compute_space import COMPUTE_SPACE_PACKAGE_DIR
 from self_host_cli.config_gen import generate_config
 
 _PID_DIR = os.path.expanduser("~/.openhost/run")
